refactor(attributes): resolve the component wire type in one place
Component::resolveType (the render base) and Column::resolveType (the column
base, added with #[AsColumn]) were near-identical reflect-and-cache copies — and
the column could not inherit the first since it does not extend the component
base. Move the resolution onto the #[Component] attribute itself as
Component::typeForClass($class), so both bases call one cached implementation.
